Fix: change top/bottom_image_height and image_gap to FLOAT to accept decimal slider values

+++ src/models/Template.js
@@ -94,17 +94,17 @@ const Template = sequelize.define('Template', {
   },
   // Top/Bottom image sizing (for two-photo-stack layout)
   top_image_height: {
-    type: DataTypes.INTEGER,
+    type: DataTypes.FLOAT,
     allowNull: true,
     defaultValue: 50, // percentage of canvas
   },
   bottom_image_height: {
-    type: DataTypes.INTEGER,
+    type: DataTypes.FLOAT,
     allowNull: true,
     defaultValue: 50, // percentage of canvas
   },
   image_gap: {
-    type: DataTypes.INTEGER,
+    type: DataTypes.FLOAT,
     allowNull: true,
     defaultValue: 0, // gap between images in pixels
   },
